requirements.txt
gTTS==2.3.1
pyttsx3==2.90
requests==2.31.0
python-dotenv==1.0.0